--- src/redux/saga/rootSaga.js
@@ -1,13 +1,0 @@
-import {all} from 'redux-saga/effects';
-import {addBookWatcher, deleteBookWatcher,
-    editBookWatcher, filterWatcher,
-    getBookByIdWatcher, updateBooksWatcher,
-} from "./watchers";
-
-//Создание корневой саги со всеми вотчерами
-function* rootSaga(){
-    yield all([updateBooksWatcher(), addBookWatcher(), deleteBookWatcher(),
-        editBookWatcher(), filterWatcher(), getBookByIdWatcher()]);
-}
-
-export {rootSaga}
