--- src/redux/saga/watchers/index.js
@@ -1,9 +1,0 @@
-import {updateBooksWatcher} from "./updateBooks";
-import {addBookWatcher} from "./addBook";
-import {deleteBookWatcher} from "./deleteBook";
-import {editBookWatcher} from "./editBook";
-import {filterWatcher} from "./filterBook";
-import {getBookByIdWatcher} from "./getBookById";
-
-export {updateBooksWatcher, addBookWatcher, deleteBookWatcher,
-    editBookWatcher, filterWatcher, getBookByIdWatcher}
